Move transfer funds logic to transfer facade
